support parallel mode for mocha
