+++ ext/exporters.d.ts
@@ -5,6 +5,7 @@ declare const exporters: {
   AGENT_PROXY: 'agent_proxy',
   JEST_WORKER: 'jest_worker',
   CUCUMBER_WORKER: 'cucumber_worker'
+  MOCHA_WORKER: 'mocha_worker'
 }
 
 export = exporters
