+++ ext/exporters.js
@@ -5,5 +5,6 @@ module.exports = {
   DATADOG: 'datadog',
   AGENT_PROXY: 'agent_proxy',
   JEST_WORKER: 'jest_worker',
-  CUCUMBER_WORKER: 'cucumber_worker'
+  CUCUMBER_WORKER: 'cucumber_worker',
+  MOCHA_WORKER: 'mocha_worker'
 }
