+++ packages/datadog-plugin-mocha/src/index.js
@@ -20,7 +20,14 @@ const {
   removeEfdStringFromTestName,
   TEST_IS_NEW,
   TEST_IS_RETRY,
-  TEST_EARLY_FLAKE_ENABLED
+  TEST_EARLY_FLAKE_ENABLED,
+  TEST_SESSION_ID,
+  TEST_MODULE_ID,
+  TEST_MODULE,
+  TEST_SUITE_ID,
+  TEST_COMMAND,
+  TEST_SUITE,
+  MOCHA_IS_PARALLEL
 } = require('../../dd-trace/src/plugins/util/test')
 const { COMPONENT } = require('../../dd-trace/src/constants')
 const {
@@ -33,6 +40,22 @@ const {
   TELEMETRY_ITR_UNSKIPPABLE,
   TELEMETRY_CODE_COVERAGE_NUM_FILES
 } = require('../../dd-trace/src/ci-visibility/telemetry')
+const id = require('../../dd-trace/src/id')
+const log = require('../../dd-trace/src/log')
+
+function getTestSuiteLevelVisibilityTags (testSuiteSpan) {
+  const testSuiteSpanContext = testSuiteSpan.context()
+  const suiteTags = {
+    [TEST_SUITE_ID]: testSuiteSpanContext.toSpanId(),
+    [TEST_SESSION_ID]: testSuiteSpanContext.toTraceId(),
+    [TEST_COMMAND]: testSuiteSpanContext._tags[TEST_COMMAND],
+    [TEST_MODULE]: 'mocha'
+  }
+  if (testSuiteSpanContext._parentId) {
+    suiteTags[TEST_MODULE_ID] = testSuiteSpanContext._parentId.toString(10)
+  }
+  return suiteTags
+}
 
 class MochaPlugin extends CiPlugin {
   static get id () {
@@ -50,7 +73,8 @@ class MochaPlugin extends CiPlugin {
       if (!this.libraryConfig?.isCodeCoverageEnabled) {
         return
       }
-      const testSuiteSpan = this._testSuites.get(suiteFile)
+      const testSuite = getTestSuitePath(suiteFile, this.sourceRoot)
+      const testSuiteSpan = this._testSuites.get(testSuite)
 
       if (!coverageFiles.length) {
         this.telemetry.count(TELEMETRY_CODE_COVERAGE_EMPTY)
@@ -73,16 +97,20 @@ class MochaPlugin extends CiPlugin {
     })
 
     this.addSub('ci:mocha:test-suite:start', ({
-      testSuite,
+      testSuiteAbsolutePath,
       isUnskippable,
       isForcedToRun,
       itrCorrelationId
     }) => {
-      const store = storage.getStore()
+      // If the test module span is undefined, the plugin has not been initialized correctly and we bail out
+      if (!this.testModuleSpan) {
+        return
+      }
+      const testSuite = getTestSuitePath(testSuiteAbsolutePath, this.sourceRoot)
       const testSuiteMetadata = getTestSuiteCommonTags(
         this.command,
         this.frameworkVersion,
-        getTestSuitePath(testSuite, this.sourceRoot),
+        testSuite,
         'mocha'
       )
       if (isUnskippable) {
@@ -109,6 +137,7 @@ class MochaPlugin extends CiPlugin {
       if (itrCorrelationId) {
         testSuiteSpan.setTag(ITR_CORRELATION_ID, itrCorrelationId)
       }
+      const store = storage.getStore()
       this.enter(testSuiteSpan, store)
       this._testSuites.set(testSuite, testSuiteSpan)
     })
@@ -140,6 +169,10 @@ class MochaPlugin extends CiPlugin {
       const span = this.startTestSpan(testInfo)
 
       this.enter(span, store)
+    })
+
+    this.addSub('ci:mocha:worker:finish', () => {
+      this.tracer._exporter.flush()
     })
 
     this.addSub('ci:mocha:test:finish', (status) => {
@@ -194,7 +227,8 @@ class MochaPlugin extends CiPlugin {
       hasForcedToRunSuites,
       hasUnskippableSuites,
       error,
-      isEarlyFlakeDetectionEnabled
+      isEarlyFlakeDetectionEnabled,
+      isParallel
     }) => {
       if (this.testSessionSpan) {
         const { isSuitesSkippingEnabled, isCodeCoverageEnabled } = this.libraryConfig || {}
@@ -204,6 +238,10 @@ class MochaPlugin extends CiPlugin {
         if (error) {
           this.testSessionSpan.setTag('error', error)
           this.testModuleSpan.setTag('error', error)
+        }
+
+        if (isParallel) {
+          this.testSessionSpan.setTag(MOCHA_IS_PARALLEL, 'true')
         }
 
         addIntelligentTestRunnerSpanTags(
@@ -234,6 +272,37 @@ class MochaPlugin extends CiPlugin {
       this.libraryConfig = null
       this.tracer._exporter.flush()
     })
+
+    this.addSub('ci:mocha:worker-report:trace', (traces) => {
+      const formattedTraces = JSON.parse(traces).map(trace =>
+        trace.map(span => {
+          const formattedSpan = {
+            ...span,
+            span_id: id(span.span_id),
+            trace_id: id(span.trace_id),
+            parent_id: id(span.parent_id)
+          }
+          if (formattedSpan.name === 'mocha.test') {
+            const testSuite = span.meta[TEST_SUITE]
+            const testSuiteSpan = this._testSuites.get(testSuite)
+            if (!testSuiteSpan) {
+              log.warn(`Test suite span not found for test span with test suite ${testSuite}`)
+              return formattedSpan
+            }
+            const suiteTags = getTestSuiteLevelVisibilityTags(testSuiteSpan)
+            formattedSpan.meta = {
+              ...formattedSpan.meta,
+              ...suiteTags
+            }
+          }
+          return formattedSpan
+        })
+      )
+
+      formattedTraces.forEach(trace => {
+        this.tracer._exporter.export(trace)
+      })
+    })
   }
 
   startTestSpan (testInfo) {
@@ -242,7 +311,8 @@ class MochaPlugin extends CiPlugin {
       title,
       isNew,
       isEfdRetry,
-      testStartLine
+      testStartLine,
+      isParallel
     } = testInfo
 
     const testName = removeEfdStringFromTestName(testInfo.testName)
@@ -257,8 +327,12 @@ class MochaPlugin extends CiPlugin {
       extraTags[TEST_SOURCE_START] = testStartLine
     }
 
+    if (isParallel) {
+      extraTags[MOCHA_IS_PARALLEL] = 'true'
+    }
+
     const testSuite = getTestSuitePath(testSuiteAbsolutePath, this.sourceRoot)
-    const testSuiteSpan = this._testSuites.get(testSuiteAbsolutePath)
+    const testSuiteSpan = this._testSuites.get(testSuite)
 
     if (this.repositoryRoot !== this.sourceRoot && !!this.repositoryRoot) {
       extraTags[TEST_SOURCE_FILE] = getTestSuitePath(testSuiteAbsolutePath, this.repositoryRoot)
