+++ packages/dd-trace/test/ci-visibility/exporters/test-worker/exporter.spec.js
@@ -6,16 +6,19 @@ const TestWorkerCiVisibilityExporter = require('../../../../src/ci-visibility/ex
 const {
   JEST_WORKER_TRACE_PAYLOAD_CODE,
   JEST_WORKER_COVERAGE_PAYLOAD_CODE,
-  CUCUMBER_WORKER_TRACE_PAYLOAD_CODE
+  CUCUMBER_WORKER_TRACE_PAYLOAD_CODE,
+  MOCHA_WORKER_TRACE_PAYLOAD_CODE
 } = require('../../../../src/plugins/util/test')
 
 describe('CI Visibility Test Worker Exporter', () => {
   let send, originalSend
+
   beforeEach(() => {
     send = sinon.spy()
     originalSend = process.send
     process.send = send
   })
+
   afterEach(() => {
     process.send = originalSend
   })
@@ -77,6 +80,31 @@ describe('CI Visibility Test Worker Exporter', () => {
       const cucumberWorkerExporter = new TestWorkerCiVisibilityExporter()
       cucumberWorkerExporter.export(trace)
       cucumberWorkerExporter.flush()
+      expect(send).not.to.have.been.called
+    })
+  })
+  context('when the process is a MOCHA worker', () => {
+    beforeEach(() => {
+      process.env.MOCHA_WORKER_ID = '1'
+    })
+    afterEach(() => {
+      delete process.env.MOCHA_WORKER_ID
+    })
+    it('can export traces', () => {
+      const trace = [{ type: 'test' }]
+      const traceSecond = [{ type: 'test', name: 'other' }]
+      const mochaWorkerExporter = new TestWorkerCiVisibilityExporter()
+      mochaWorkerExporter.export(trace)
+      mochaWorkerExporter.export(traceSecond)
+      mochaWorkerExporter.flush()
+      expect(send).to.have.been.calledWith([MOCHA_WORKER_TRACE_PAYLOAD_CODE, JSON.stringify([trace, traceSecond])])
+    })
+    it('does not break if process.send is undefined', () => {
+      delete process.send
+      const trace = [{ type: 'test' }]
+      const mochaWorkerExporter = new TestWorkerCiVisibilityExporter()
+      mochaWorkerExporter.export(trace)
+      mochaWorkerExporter.flush()
       expect(send).not.to.have.been.called
     })
   })
